+++ CHANGELOG.md
@@ -12,3 +12,10 @@ All commits will be recorded here.
 ## [1.0.1] `file_browser` and `read_source` functions in utils.py
 
 - Added `file_browser` and `read_source` functions to utils.py
+
+---
+
+# [1.0.2] `metadata.py` functions including EXIF extraction
+
+- Added functions which return EXIF data from a photo
+- Added dms to decimal conversion to `utils.py`